Change ZeroCoordinate to plain number
This type is probably not worth having to cast number to ZeroCoordinate
all over the place.

Also, change OneCoordinate to OneIndexedCoordinate. (I kept this one
because it's probably not going to be used anywhere except SquareName,
though maybe I'll end up changing SquareName to string to avoid similar
casts. We'll see...)